+++ packages/backend/src/models/DashboardModel/DashboardModel.mock.ts
@@ -289,6 +289,7 @@ export const user: SessionUser = {
     lastName: 'lastName',
     organizationUuid: 'organizationUuid',
     organizationName: 'organizationName',
+    organizationCreatedAt: new Date(),
     isTrackingAnonymized: false,
     isMarketingOptedIn: false,
     isSetupComplete: true,